// src2/libs/imng-oidc-client/src/lib/models/oidc-user.ts
/* eslint-disable @typescript-eslint/no-explicit-any */
export interface IOidcUser {
  id_token: string;
  session_state?: any;
  access_token: string;
  refresh_token?: string;
  token_type: string;
  scope: string;
  profile: any;
  expires_at: number;
  state?: string;
  toStorageString?: () => string;
  readonly expires_in?: number;
  readonly expired?: boolean;
  readonly scopes?: string[];
}